chore: move utils specs to specs folder
